requirements.txt
fastapi
uvicorn
motor
PyJWT
bcrypt
dotenv